+++ ChangeLog.md
@@ -6,9 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Added
+
+- added dirty read support ([reading from followers](https://docs.arangodb.com/current/Manual/Administration/ActiveFailover/#reading-from-follower)) for AQL queries
+  - added `QueryOptions#allowDirtRead`
+
 ### Changed
 
 - use AQL `LIKE` instead of `REGEX_TEST` for query by example in `ArangoRepository`
+- upgraded dependency arangodb-java-driver 5.0.0
 
 ## [3.0.0] - 2018-09-04